Remove fallbackEnvironmentResolver checks and move logic to SchemaProvider
